# docs/sharepoint_sim_openapi.json
{"openapi":"3.1.0","info":{"title":"SharePoint CSV Simulator API","version":"0.1.0"},"paths":{"/sim/generate":{"post":{"tags":["sim"],"summary":"Generate","description":"Generate one or more datasets by name.\n\nArgs:\n    types (str): Comma-separated dataset names (e.g., \"ACQ,Productivity\").\n    rows (int, optional): Row count override for all datasets. Defaults to None.\n\nReturns:\n    dict: Metadata for created files. Example::\n        {\n            \"files\": [\n                {\"filename\": \"ACQ__2025-08-18_0200.csv\", \"size\": 1234},\n                ...\n            ]\n        }\n\nRaises:\n    HTTPException: If any dataset name is unknown (400).","operationId":"generate_sim_generate_post","parameters":[{"name":"types","in":"query","required":true,"schema":{"type":"string","title":"Types"}},{"name":"rows","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rows"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"array","items":{"type":"object","additionalProperties":{"anyOf":[{"type":"integer"},{"type":"string"}]}}},"title":"Response Generate Sim Generate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sim/generate/all":{"post":{"tags":["sim"],"summary":"Generate All","description":"Generate all known datasets in a single call.\n\nArgs:\n    rows (int, optional): Row count override for all datasets. Defaults to None.\n\nReturns:\n    dict: Metadata for created files. Example::\n        {\n            \"files\": [\n                {\"filename\": \"ACQ__2025-08-18_0200.csv\", \"size\": 1234},\n                ...\n            ]\n        }","operationId":"generate_all_sim_generate_all_post","parameters":[{"name":"rows","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rows"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"array","items":{"type":"object","additionalProperties":{"anyOf":[{"type":"integer"},{"type":"string"}]}}},"title":"Response Generate All Sim Generate All Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sim/files":{"get":{"tags":["sim"],"summary":"Files","description":"List all generated simulator CSV files.\n\nReturns:\n    dict: List of file metadata (filename, size). Example::\n        {\n            \"files\": [\n                {\"filename\": \"ACQ__2025-08-18_0200.csv\", \"size\": 1234},\n                ...\n            ]\n        }","operationId":"files_sim_files_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"items":{"additionalProperties":{"anyOf":[{"type":"integer"},{"type":"string"}]},"type":"object"},"type":"array"},"type":"object","title":"Response Files Sim Files Get"}}}}}}},"/sim/datasets":{"get":{"tags":["sim"],"summary":"List Datasets","description":"Return metadata for each dataset.\n\nReturns:\n    dict: List of dataset descriptors with keys: name, headers, roles. Example::\n        {\n            \"datasets\": [\n                {\"name\": \"ACQ\", \"headers\": [...], \"roles\": [...]},\n                ...\n            ]\n        }","operationId":"list_datasets_sim_datasets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"items":{"additionalProperties":true,"type":"object"},"type":"array"},"type":"object","title":"Response List Datasets Sim Datasets Get"}}}}}}},"/sim/datasets/{name}":{"get":{"tags":["sim"],"summary":"Get Dataset","description":"Return schema information for a single dataset.\n\nArgs:\n    name (str): Dataset name (e.g., \"ACQ\").\n\nReturns:\n    dict: Dataset descriptor with keys: name, headers, roles.\n\nRaises:\n    HTTPException: If dataset is unknown (404).","operationId":"get_dataset_sim_datasets__name__get","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Dataset Sim Datasets  Name  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sim/spec":{"get":{"tags":["sim"],"summary":"Spec Addendum","description":"Return the simulator spec addendum markdown (or placeholder).\n\nReturns:\n    str: Markdown content of the spec addendum, or a not found message.","operationId":"spec_addendum_sim_spec_get","responses":{"200":{"description":"Successful Response","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/sim/download/{filename}":{"get":{"tags":["sim"],"summary":"Download","description":"Download a generated CSV file by filename.\n\nArgs:\n    filename (str): Name of the file to download.\n\nReturns:\n    Response: CSV file contents with text/csv media type.\n\nRaises:\n    HTTPException: If file is not found (404).","operationId":"download_sim_download__filename__get","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sim/reset":{"post":{"tags":["sim"],"summary":"Reset","description":"Delete all generated simulator files and reset state.\n\nReturns:\n    Response: Empty response with status 204.","operationId":"reset_sim_reset_post","responses":{"204":{"description":"Successful Response"}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}